+++ CHANGELOG.md
@@ -138,7 +138,7 @@ All notable changes to this project will be documented in this file - [docs/chan
 - OpenTracing compliance tha can be used for manual instrumentation
 
 [Unreleased]: https://github.com/DataDog/dd-trace-php/compare/0.5.1...HEAD
-[0.5.0]: https://github.com/DataDog/dd-trace-php/compare/0.5.0...0.5.1
+[0.5.1]: https://github.com/DataDog/dd-trace-php/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/DataDog/dd-trace-php/compare/0.4.2...0.5.0
 [0.4.2]: https://github.com/DataDog/dd-trace-php/compare/0.4.1...0.4.2
 [0.4.1]: https://github.com/DataDog/dd-trace-php/compare/0.4.0...0.4.1